model/api/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
keras>=3.0.0
tensorflow-cpu>=2.15.0
pillow>=10.1.0
numpy>=1.24.3
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
setuptools>=68.0.0
wheel>=0.40.0
